rename sub-namespace `FromPackageJson` -> `FromNodePackageJson`
