+++ src/factories/fromNodePackageJson.node.ts
